requirements.txt
# ============================================================
# PYTHON DEPENDENCIES
# ============================================================
# Install with: pip install -r requirements.txt
# ============================================================

# ============================================================
# CORE DEPENDENCIES
# ============================================================

# Environment variable management
python-dotenv>=1.0.0

# Cross-platform colored terminal output
colorama>=0.4.6

# ============================================================
# DATABASE (PostgreSQL)
# ============================================================

psycopg2-binary>=2.9.0

# ============================================================
# EXCEL PROCESSING
# ============================================================

openpyxl>=3.1.0

# ============================================================
# BROWSER AUTOMATION (Downloaders)
# ============================================================

playwright>=1.40.0

# ============================================================
# TELEGRAM ALERTS
# ============================================================

requests>=2.31.0

# ============================================================
# STREAMLIT UI (Future)
# ============================================================
# Will be added when UI is implemented
# streamlit==1.29.0

# ============================================================
# TESTING (Future)
# ============================================================
# Will be added when tests are written
# pytest==7.4.3
# pytest-cov==4.1.0
